Backend/requirements.txt
# Core Framework
# FastAPI + Pydantic v2 (recommended for Python 3.12)
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.23
alembic==1.12.1
aiosqlite==0.20.0  # SQLite async driver (default for local dev)

# Chess Engine & Analysis
python-chess==1.999
stockfish==3.28.0

# API Clients
httpx==0.25.2
# aiohttp removed for lighter, 3.14-friendly install

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Utilities
python-dotenv==1.0.0
tenacity==8.2.3  # Retry logic

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
mypy==1.7.1
